requirements.txt
python-dotenv
streamlit
langchain-community
fastapi
uvicorn
sse_starlette
pydantic
langchain-core
langserve
langchain-mistralai
mistralai
chromadb
faiss-cpu
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
langchain-google-community
langgraph
langchain-googledrive